// frontend/src/libs/api/generated/orval/model/bandRecruitment.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Muslog API
 * OpenAPI spec version: 1.0.0
 */

export interface BandRecruitment {
  id?: string;
  title?: string;
  description?: string;
  genre?: string;
  location?: string;
  recruitingParts?: string[];
  skillLevel?: string;
  contact?: string;
  /** @nullable */
  deadline?: string | null;
  status?: string;
  userId?: string;
  applicationsCount?: number;
  hasApplied?: boolean;
  createdAt?: string;
  updatedAt?: string;
}